pyproject.toml
[tool.poetry]
name = "teleminer"
version = "0.1.0"
description = ""
authors = ["fauzaanu <fauzaangasim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tgcrypto = "^1.2.5"
python-dotenv = "^1.0.1"
pyrofork = "^2.3.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
